added evals template | updated run_evals skript to work with all planners

+++ eval/src/run_evals.py
@@ -1,4 +1,4 @@
-from typing import Tuple, Union
+from typing import Any, Union
 import rospy
 import rospkg
 import yaml
@@ -12,85 +12,86 @@ import argparse
 __all__ = ["main", "evaluate_scenario"]
 
 
-""" GLOBAL PARAMS """
-ROSNAV_3D_BASEPATH = os.path.join(rospkg.RosPack().get_path("eval"), "..")
-BUILD_STARTUP_CMD = (
-    lambda use_recorder, trainings_env, network_type, robot, world, scenario: "roslaunch arena_bringup "
-    f"start_arena_gazebo.launch use_recorder:={use_recorder} use_rviz:=false gui:=false trainings_environment"
-    f"trainings_environment:={trainings_env} network_type:={network_type} model:={robot} "
-    f"local_planner:=rosnav world:={world} scenario_file:={scenario}"
-)
-DOCKER_CMD = (
-    lambda launch_docker: f"{ROSNAV_3D_BASEPATH}/docker/drl_agent_node/start_docker.sh {ROSNAV_3D_BASEPATH} {launch_docker}"
-)
-
-
-def read_evals_file(file_name: str) -> Union[Tuple[str, str, str, str], None]:
-    file_path = os.path.join(
-        rospkg.RosPack().get_path("eval"), "src", f"{file_name}.yaml"
-    )
-    with open(file_path, "r") as file:
-        evals = yaml.safe_load(file)
-
-        should_launch_docker = evals["shouldLaunchDocker"]
-
-        if should_launch_docker:
-            dockerName = evals["dockerName"]
-
-        return (
-            evals["trainings_env"],
-            evals["network_type"],
-            evals["scenarios"],
-            dockerName if should_launch_docker else None,
+class EvalDeclerationFile:
+    def __init__(
+        self, 
+        local_planner: str, 
+        additional_args: Union[Any, None], # None or a List with dicts containing "value": str, "name": str
+        scenarios: Any, # Array of Dict with entries "robot": str and worlds: Array of Dict with "world": str, "scenario": str
+        docker_name: Union[None, str]
+    ) -> None:
+        self.local_planner = local_planner
+        self.arguments = additional_args
+        self.scenarios = scenarios
+        self.docker_name = docker_name
+    
+    def get_startup_command_creator(self, use_recorder: bool, use_rviz: bool, verbose: bool) -> str:
+        return (lambda robot, world, scenario:
+            f"roslaunch arena_bringup start_arena_gazebo.launch use_recorder:={use_recorder} gui:=false "
+            f"use_rviz:={use_rviz} {self.build_arguments_string()} model:={robot} "
+            f"scenario_file:={scenario} world:={world} local_planner:={self.local_planner}"
+            + (f" > eval_{self.local_planner}.txt" if verbose else "") 
         )
+
+    def build_arguments_string(self) -> str:
+        if not self.arguments:
+            return ""
+
+        return " ".join([f"{arg['name']}:={arg['value']}" for arg in self.arguments])
+
+    def build_docker_command(self) -> Union[str, None]:
+        if not self.docker_name:
+            return None
+        
+        rosnav_3d_path = os.path.join(rospkg.RosPack().get_path("eval"), "..")
+        return f"{rosnav_3d_path}/docker/drl_agent_node/start_docker.sh {rosnav_3d_path} {self.docker_name}"
+        
+    @staticmethod
+    def read_eval_file(file_name: str):
+        file_path = os.path.join(
+            rospkg.RosPack().get_path("eval"), "src", f"{file_name}.yaml"
+        )
+
+        def try_read_key(file, key):
+            try:
+                return file[key]
+            except:
+                return None
+
+        with open(file_path, "r") as file:
+            evals = yaml.safe_load(file)
+
+            additional_arguments = try_read_key(evals, "additionalArguments")
+            docker_name = try_read_key(evals, "dockerName")
+
+            return EvalDeclerationFile(
+                evals["local_planner"],
+                additional_arguments,
+                evals["scenarios"],
+                docker_name
+            )
 
 
 def evaluate_scenario(
-    scenario_file: str,
-    current_world: str,
-    trainings_env: str,
-    network_type: str,
-    robot: str,
-    launch_docker: bool,
-    use_recorder: bool,
-    verbose: bool = False,
+    startup_command: str,
+    docker_command: Union[str, None]
 ) -> None:
     """_summary_
 
     Args:
-        scenario_file (str): _description_
-        current_world (str): _description_
-        use_recorder (bool): _description_
-        trainings_env (str): _description_
-        network_type (str): _description_
-        robot (str): _description_
-        launch_docker (bool): _description_
+        startup_command (str): _description_
+        docker_command (Union[str, None]): _description_ 
     """
-    print("===============================================================")
-    print(
-        f"[{robot}] Starting scenario {scenario_file} in world {current_world}"
-    )
-
-    startup_command = BUILD_STARTUP_CMD(
-        use_recorder,
-        trainings_env,
-        network_type,
-        robot,
-        current_world,
-        scenario_file,
-    )
-    if verbose:
-        startup_command += "> eval_log.txt"
 
     process = subprocess.Popen(
         startup_command, shell=True, preexec_fn=os.setsid
     )
 
-    if launch_docker:
-        time.sleep(2)
+    if docker_command:
+        time.sleep(5)
         print("Launching docker")
         docker_process = subprocess.Popen(
-            DOCKER_CMD(launch_docker), shell=True, preexec_fn=os.setsid
+            docker_command, shell=True, preexec_fn=os.setsid
         )
 
     time.sleep(1)
@@ -99,36 +100,40 @@ def evaluate_scenario(
     rospy.wait_for_message("/End_of_scenario", Bool)
 
     os.killpg(os.getpgid(process.pid), signal.SIGINT)
-    os.killpg(os.getpgid(docker_process.pid), signal.SIGINT)
-
     time.sleep(1)
-
     os.system("killall roslaunch")
-
-    if launch_docker:
-        os.system("killall docker")
-        docker_process.wait(10)
 
     process.wait(10)
 
-    time.sleep(10)
+    if docker_command:
+        os.killpg(os.getpgid(docker_process.pid), signal.SIGINT)
+        time.sleep(1)
+        os.system("killall docker")
 
-    print("===============================================================")
-    print(
-        f"[{robot}] Scenario {scenario_file} in world {current_world} finished"
-    )
+        docker_process.wait(10)
+
+    time.sleep(10)
 
 
 def main(args):
     use_recorder = args.use_recorder
     eval_file = args.eval_file
     verbose = args.verbose
+    use_rviz = args.use_rviz
 
-    trainings_env, network_type, scenarios, launch_docker = read_evals_file(
-        eval_file
+    eval_decleration_file = EvalDeclerationFile.read_eval_file(eval_file)
+
+    if not eval_decleration_file:
+        raise Exception()
+
+    docker_command = eval_decleration_file.build_docker_command()
+    build_startup_command = eval_decleration_file.get_startup_command_creator(
+        use_recorder, 
+        use_rviz, 
+        verbose
     )
 
-    for scenario in scenarios:
+    for scenario in eval_decleration_file.scenarios:
         robot = scenario["robot"]
         worlds = scenario["worlds"]
 
@@ -136,17 +141,21 @@ def main(args):
             current_world = world["world"]
             scenario_file = world["scenario"]
 
-            evaluate_scenario(
-                scenario_file,
-                current_world,
-                trainings_env,
-                network_type,
-                robot,
-                launch_docker,
-                use_recorder,
-                verbose,
+            startup_command = build_startup_command(robot, current_world, scenario_file)
+
+            print("===============================================================")
+            print(
+                f"[{robot}] Starting scenario {scenario_file} in world {current_world}"
             )
 
+            print(startup_command)
+    
+            evaluate_scenario(startup_command, docker_command)
+
+            print("===============================================================")
+            print(
+                f"[{robot}] Scenario {scenario_file} in world {current_world} finished"
+            )
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
@@ -154,5 +163,9 @@ if __name__ == "__main__":
     parser.add_argument("-f", "--eval_file")
     parser.add_argument("-r", "--use_recorder", action="store_true")
     parser.add_argument("-v", "--verbose", action="store_true")
+    parser.add_argument("--use_rviz", action="store_true")
 
-    main(parser.parse_args())
+    try: 
+        main(parser.parse_args())
+    except:
+        os.system("killall roslaunch")
